requirements.txt
fastapi
uvicorn
scikit-learn
numpy
pandas
requests
pillow
python-multipart
